requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
langchain==0.1.1
langchain-google-genai==0.0.11
langchain-core==0.1.13
google-generativeai==0.3.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0